Use EntityID type and move payload construction to createEntityStatePayload
